+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - Added `modify` and `modify_mut` methods to `UniqueMmioPointer<ReadWrite<T>>` and
   `UniqueMmioPointer<ReadPureWrite<T>>`.
+- Added `get_range()` methods to `UniqueMmioPointer<[T]>`, `UniqueMmioPointer<[T; LEN]>`,
+  `SharedMmioPointer<[T]>`, `SharedMmioPointer<[T; LEN]>`. These methods return a pointer to the
+  subset of elements, specified by a `Range`.
+- Added `iter()` methods to `UniqueMmioPointer<[T]>`, `UniqueMmioPointer<[T; LEN]>`,
+  `SharedMmioPointer<[T]>`, `SharedMmioPointer<[T; LEN]>`. These methods return an iterator to the
+  items of the pointer slices or arrays.
 
 ## 0.2.7